// src/app/interfaces/post-data.ts
export interface PostData {
    action: string;
    status: string;
    id: string;
    comment: string;
    reason: string;
    goods: string;
    payment: string;
    new_plam_date: string;
    check: string;
    recognizedCheckData: string;
    sign: string;
    cardNums: string;
    waitingMinutes: number;
}